+++ clang-tidy/btc/UnmanagedDerivedFromSharableCheck.h
@@ -1,0 +1,38 @@
+//===--- UnmanagedDerivedFromSharableCheck.h -
+//clang-tidy--------------------------*- C++ -*-===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BTC_DERIVEDFROMSHARABLESTACKPROHIBITOIONCHECKCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BTC_DERIVEDFROMSHARABLESTACKPROHIBITOIONCHECKCHECK_H
+
+#include "../ClangTidy.h"
+
+namespace clang {
+namespace tidy {
+namespace btc {
+
+/// Internal Btc-ag check. Purpose is to safeguard from creating Sharable based
+/// objects on stack.
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/btc-unmanaged-derived-from-sharable.html
+class UnmanagedDerivedFromSharableCheck : public ClangTidyCheck {
+ public:
+  UnmanagedDerivedFromSharableCheck(StringRef Name,
+                                           ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+};
+
+}  // namespace btc
+}  // namespace tidy
+}  // namespace clang
+
+#endif  // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BTC_DERIVEDFROMSHARABLESTACKPROHIBITOIONCHECKCHECK_H