+++ clang-tidy/tool/ClangTidyMain.cpp
@@ -494,6 +494,11 @@ extern volatile int BoostModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED BoostModuleAnchorDestination =
     BoostModuleAnchorSource;
 
+// This anchor is used to force the linker to link the BtcModule.
+extern volatile int BtcModuleAnchorSource;
+static int LLVM_ATTRIBUTE_UNUSED BtcModuleAnchorDestination =
+    BtcModuleAnchorSource;
+
 // This anchor is used to force the linker to link the BugproneModule.
 extern volatile int BugproneModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED BugproneModuleAnchorDestination =